⚡️ Update perpormance

+++ controller/users.js
@@ -132,6 +132,15 @@ exports.removeClient = asyncHandler(async (req, res, next) => {
         userId: god.checker_id,
       },
     });
+
+    const blank = await req.db.blank.findOne({
+      where: {
+        clientId: clientId,
+      },
+    });
+    if (blank) {
+      throw new MyError("Устгах боломжгүй байна!!");
+    }
     await client.destroy();
   } else {
     const client = await req.db.clients.findOne(req.body, {
@@ -140,6 +149,15 @@ exports.removeClient = asyncHandler(async (req, res, next) => {
         userId,
       },
     });
+
+    const blank = await req.db.blank.findOne({
+      where: {
+        clientId: clientId,
+      },
+    });
+    if (blank) {
+      throw new MyError("Устгах боломжгүй байна!!");
+    }
     await client.destroy();
   }
 
@@ -323,7 +341,7 @@ exports.signup = asyncHandler(async (req, res, next) => {
 
   const message = `<b>Сайн байна уу?</b><br>
   Та “Хэвлэмэл хуудасны захиалга, хяналтын систем”-д дараах эрхээр нэвтрэн орж хэвлэмэл хуудасны захиалгаа өгнө үү.<br>
-  Холбоос: <a href="www.eblank.mn">www.eblank.mn</a><br>
+  Холбоос: <a href="dashboard.eblank.mn">dashboard.eblank.mn</a><br>
   Нэвтрэх нэр: <b>${user.phone_number}</b><br>
   Нууц үг: <b>${password}</b><br>
   Өдрийг сайхан өнгөрүүлээрэй!<br>
